Allow passing wp download

+++ add_site.sh
@@ -20,11 +20,13 @@ echo DB_PASSWORD=$PW> pw/$site
 docker run --name gendb -d -e MYSQL_ROOT_PASSWORD=$MYSQL_ROOT_PASSWORD -v $PWD/db:/var/lib/mysql -v $PWD/sql/$site.sql:/docker-entrypoint-initdb.d/schema.sql mariadb
 cd sites/$site
 
-wget https://fi.wordpress.org/latest-fi.tar.gz
-tar xvfz latest-fi.tar.gz
-rm *.gz
-mv wordpress/* .
-rmdir wordpress
+if [ "$1" != "nowp" ]; then
+	wget https://fi.wordpress.org/latest-fi.tar.gz
+	tar xvfz latest-fi.tar.gz
+	rm *.gz
+	mv wordpress/* .
+	rmdir wordpress
+fi
 sleep 5
 docker stop gendb
 docker rm gendb
